"added file for edges, finished build graph, working on insert and edges"

+++ Edge.h
@@ -10,7 +10,7 @@ private:
     Vertex<class Data> *end;
 
 public:
-    Edge(Vertex<class Data> *begin, Vertex<class Data> *end, int weight){
+    Edge(Vertex<class Data> *begin, Vertex<class Data> *end, double weight){
         this->begin = begin;
         this->end = end;
         this->weight = weight;
